libion: Do not print error logs for unimplemented IOCTLs
Various versions of ION are supported by libion. Inorder to check which
version is supported by the device, libion issues various IOCTLs to
check whether they are supported by the device or not. Based on whether
the device supports the IOCTL, libion determines the version of ION
supported by the device. Although the error log is printed only once per
process, it could cause confusion during debugging. Hence, this patch
does not print the error log when the errno returned is ENOTTY which
indicates that an IOCTL is unsupported by a device.

Test: manual
Bug: 188388236
Change-Id: I2df9b1d0c71577f406434540e883f71c6796d584